requirements.txt
cycler==0.10.0
decorator==4.4.1
fastapi==0.48.0
h11==0.9.0
httptools==0.0.13
imageio==2.6.1
imgaug==0.4.0
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pkg-resources==0.0.0
pydantic==1.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-multipart==0.0.5
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
starlette==0.12.9
torch==1.4.0+cpu
torchvision==0.5.0+cpu
uvicorn==0.11.2
uvloop==0.14.0
websockets==8.1
